requirements.txt
click==7.0
PyJWT==1.7.1
requests==2.22.0
